chore: refactor and types
