+++ src/utils/boardIdUnserializer.ts
@@ -1,6 +1,12 @@
 import { makeMarshal } from '@endo/marshal';
+import type { Marshal } from '@endo/marshal';
+
+type IdentityMarshal = Marshal<string>;
 
 /**
  *  Unserializes values from the board as just their ID strings.
  */
-export const boardIdUnserializer = makeMarshal(undefined, slot => slot);
+export const identityMarshal: IdentityMarshal = makeMarshal(
+  undefined,
+  slot => slot
+);
